refactor(utils): split `platform` into native/web modules
Replaces the inline `#[cfg]`-gated `mod imp` blocks in `platform.rs`
with `#[path]`-mounted `platform/native.rs` and `platform/web.rs`,
each cfg-free.
