+++ utils/src/platform/web.rs
@@ -1,0 +1,5 @@
+pub trait MaybeSend {}
+impl<T> MaybeSend for T {}
+
+pub trait MaybeSendSync: MaybeSend {}
+impl<T> MaybeSendSync for T {}
